Revert "Others fixes and small patches (various)"
This reverts commit b18752d407f9656d85a4cee2c501927a547be91c.

+++ drivers/media/video/gspca/gspca.c
@@ -1505,7 +1505,7 @@ static int vidioc_reqbufs(struct file *file, void *priv,
 	struct gspca_dev *gspca_dev = priv;
 	int i, ret = 0, streaming;
 
-	switch ((int) rb->memory) {
+	switch (rb->memory) {
 	case GSPCA_MEMORY_READ:			/* (internal call) */
 	case V4L2_MEMORY_MMAP:
 	case V4L2_MEMORY_USERPTR:
